requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
channels==4.2.0
channels_redis==4.2.1
daphne==4.1.2
celery==5.4.0
redis==5.2.1
mindee==4.35.1
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
openai>=1.82.0
rapidfuzz>=3.0.0
resend>=2.0.0
stripe>=10.0.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.160.0
